Added keypress event outside of text input
A keypress still is invoked out of the input. Using mousetrap.js, it cannot invoke its events without the keypress from occurring. Tested on Mac in Chrome. If you would like to see that the "keypress" event does invoke...

document.addEventListener('keypress', function(){
  console.log('keypress');
}, true)

This fires whether in or out of a text input.